requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
cffi==1.15.1
Django==4.1.7
django-bootstrap-v5==1.0.11
django-bootstrap3==22.2
django-braces==1.15.0
misaka==2.1.1
pycparser==2.21
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.2
